feat: hardening de seguridad, password reset y mejoras de UX
Seguridad:
- Rate limiting en /auth/login (5/min), /refresh (10/min), /forgot-password (3/min)
- Structured logging JSON con X-Request-ID en cada response
- Verificacion defensiva en background task de extraccion

Password reset completo:
- Modelo PasswordResetToken + migracion Alembic 0006
- Servicio de email (SMTP opcional; fallback a log si no configurado)
- Endpoints POST /auth/forgot-password y /auth/reset-password
- UI en login: enlace, panel de email y formulario de nueva contrasena

Fixes:
- temperature casteado a float (antes int truncaba 0.7 a 0)
- LLM_RETRY_ATTEMPTS capado a max 10 para evitar DoS por config
- Fallos de borrado de storage ahora se loguean en lugar de silenciarse
- Flash de contenido al cargar con sesion expirada corregido (auth-loading)

Historial unificado:
- Nuevo endpoint GET /assessments/runs para listar todas las ejecuciones de la BU
- Historial mezcla extracciones y assessment runs ordenados por fecha

Favicon:
- Rediseno: escudo + ojo en brand purple (coherente con el UI)

Resaltado PDF (lupa):
- Busqueda exacta por subcadena sobre texto concatenado (antes ventana de 100 items)
- Fallback fuzzy con ventana de 20 items y umbral 65% (antes 30%)

Razonamiento LLM:
- Prompt ampliado a 3-5 frases con trazabilidad: evidencia, seccion, ambiguedad y descarte

+++ app/routers/prompt_configs.py
@@ -117,7 +117,7 @@ async def create_prompt_config(
             base_prompt=data.base_prompt,
             variables=variables_dicts,
             model=data.model,
-            temperature=int(data.temperature),
+            temperature=float(data.temperature),
         )
 
         db.add(prompt_config)
